Create a NetCDFDataset super class to apply to Rgrid, Sgrid, and Ugrid
