+++ wms/models/__init__.py
@@ -5,6 +5,7 @@ from wms.models.server import Server
 from wms.models.variable import Variable
 from wms.models.layer import Layer, VirtualLayer
 
+from wms.models.datasets.netcdf import NetCDFDataset
 from wms.models.datasets.base import Dataset
 from wms.models.datasets.ugrid import UGridDataset
 from wms.models.datasets.sgrid import SGridDataset
